pyproject.toml
[tool.poetry]
name = "imgur-upload"
version = "0.1.0"
description = "The CLI tool that instantly uploads images saved on the clipboard to imgur."
authors = ["ibuibu <ibuibu69@gmail.com>"]
license = "ライセンス名"
homepage = "https://ibulog.com"
repository = "https://github.com/ibuibu/imgur-upload"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^8.3.2"
python-dotenv = "^0.19.0"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
